Add files via upload
The files in the directory GapFillingFwBw are for the implementation of the Forward-Backward Gap Filling method used initially to replace the artifact of transcranial magnetic stimulation (TMS) as registered in electroencephalogram (EEG) recordings. See the publications
1. V. K. Kimiskidis, D. Kugiumtzis, S. Papagiannopoulos, N. Vlaikidis, "Transcranial Magnetic Stimulation (TMS) Modulates Epileptiform Discharges in Patients with Frontal Lobe Epilepsy: a Preliminary EEG–TMS Study", International Journal of Neural Systems, Vol 23, 1250035, 2013
2. P. Vafidis, V.K. Kimiskidis, D. Kugiumtzis,  "Evaluation of algorithms for correction of transcranial magnetic stimulation-induced artifacts in electroencephalograms", Medical & Biological Engineering & Computing, pp 1-17, 2019
In the first paper, the method is introduced and in the second paper is compared favorably to other TMS artifact rejection methods. 
Note that the method can be used for filling gaps in any univariate time series setting.